Extract cr_table
