+++ prboom2/src/dsda/cr_table.h
@@ -1,0 +1,25 @@
+//
+// Copyright(C) 2023 by Ryan Krafnick
+//
+// This program is free software; you can redistribute it and/or
+// modify it under the terms of the GNU General Public License
+// as published by the Free Software Foundation; either version 2
+// of the License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// DESCRIPTION:
+//	DSDA CR Table
+//
+
+#ifndef __DSDA_CR_TABLE__
+#define __DSDA_CR_TABLE__
+
+#include "doomtype.h"
+
+byte* dsda_GenerateCRTable(void);
+
+#endif
